feat: tradingDetail function add controller
